moved headers, added readme_gpu
