+++ src/srv/srv_loop.c
@@ -6,7 +6,7 @@
 /*   By: kcharla <kcharla@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/31 20:52:27 by kcharla           #+#    #+#             */
-/*   Updated: 2020/11/03 23:36:08 by kcharla          ###   ########.fr       */
+/*   Updated: 2020/11/04 19:42:31 by kcharla          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
